+++ lib/src/langues/cerbere_langue.dart
@@ -1,4 +1,8 @@
+/// Langues supportées par le package Cerbère pour la localisation.
 enum CerbereLangue {
+  /// Français
   fr,
+
+  /// Anglais
   en,
 }